Removed inherited active styles and non-links that behave like links in the menu.
This change moves the active class to the link itself so menu items below aren't
affected by the active markers anymore. Additionally, items that aren't links
use a <strong>-element now. It has to be seen if this should rather be replaced
by a <span> or at least if a css rule has to be added that removes the bold
font.


git-svn-id: svn://svn.paxle.net/paxle/trunk@2044 f5d3ab42-bacf-4b1a-9692-565be65a0cfe