Added some CSS fixes for internet explorer
Using the technique described at
http://www.positioniseverything.net/easyclearing.html the header and wrapper
have been modified, this prevents - among others - the guillotine bug, see
http://www.positioniseverything.net/explorer/guillotine.html for details. The
menu works better now, too, the margin of .subnav hadn't been set to 0.
A fix for ie6 and the navi is included, too, though it doesn't seem to be
necessary for the webpages of the peer itself.


git-svn-id: svn://svn.paxle.net/paxle/trunk@2262 f5d3ab42-bacf-4b1a-9692-565be65a0cfe